1. 合并整理重复的插件部分代码片段 2. 删除无用的代码
