--- core/src/extension/common/mod.rs
@@ -1,1 +1,0 @@
-pub mod common;
