--- core/src/extension/jar/mod.rs
@@ -1,2 +1,0 @@
-mod loader;
-pub mod interface;
