+++ core/src/extension/mod.rs
@@ -1,5 +1,4 @@
-pub mod interface;
-pub mod common;
-mod jar;
-mod py;
-mod service;
+pub mod core;
+pub mod agent;
+pub mod service;
+pub mod exec;
