--- core/src/extension/py/mod.rs
@@ -1,1 +1,0 @@
-pub mod interface;
