+++ example/lib/agent-oa/extension.json
@@ -1,0 +1,9 @@
+{
+  "name": "agent-oa",
+  "version": "0.0.1",
+  "engine": "0.2.0",
+  "author": "NJ Labs",
+  "extension-requirement": [
+    "simx-http"
+  ]
+}
