--- example/lib/simx-http/extension.json
@@ -1,21 +1,0 @@
-{
-  "name": "simx-http",
-  "version": "0.0.1",
-  "engine": "0.2.0",
-  "author": "NJ Labs",
-  "dependencies": [],
-  "entry_lib": "simx-http",
-  "libs": [],
-  "init": "init",
-  "destroy": "destroy",
-  "handle_func": "handle_func",
-  "handle_service": "handle_service",
-  "service": {
-    "entry": "",
-    "init": ""
-  },
-  "func": {
-    "entry": "",
-    "init": ""
-  }
-}
